CreatedAt and UpdatedAt are converted into float64

+++ kong/ca_certificate.go
@@ -5,6 +5,6 @@ package kong
 type CACertificate struct {
 	ID        *string   `json:"id,omitempty" yaml:"id,omitempty"`
 	Cert      *string   `json:"cert,omitempty" yaml:"cert,omitempty"`
-	CreatedAt *int64    `json:"created_at,omitempty" yaml:"created_at,omitempty"`
+	CreatedAt *float64  `json:"created_at,omitempty" yaml:"created_at,omitempty"`
 	Tags      []*string `json:"tags,omitempty" yaml:"tags,omitempty"`
 }
